sketch a store handle (#263)
